이솔 -BasketVO 자료형 오류 해결 (03/20)

+++ app/src/main/java/com/example/smcart/BasketActivity.kt
@@ -64,7 +64,7 @@ class BasketActivity : AppCompatActivity() {
                     var price = jsonArray.getJSONObject(i).getString("prod_price")
                     var cnt = "1"
                     var img = jsonArray.getJSONObject(i).getString("prod_img")
-                    data.add(BasketVO(name, price.toInt(),cnt , img.toInt())).toString()
+                    data.add(BasketVO(name, price.toInt(),cnt.toInt() , img.toInt())).toString()
                     Log.d("data",data[i].toString())
                 }
